HotellingsT2 {ICSNP}	R Documentation

Hotelling's T2 Test

Description

Hotelling's T2 test for the one and two sample case.

Usage

 HotellingsT2(X, ...)

## Default S3 method:
HotellingsT2(X, Y = NULL, mu = NULL, test = "f",
             na.action = na.fail, ...)

## S3 method for class 'formula'
HotellingsT2(formula, na.action = na.fail, ...) 

Arguments

X
a numeric data frame or matrix.

Y
an optional numeric data frame or matrix for the two sample test. If NULL a one sample test is performed.

mu
a vector indicating the hypothesized value of the mean (or difference in means if a two sample test is performed). NULL represents origin or no difference between the groups.

test
if 'f', the decision is based on the F-distribution, if 'chi' a chi-squared approximation is used.

formula
a formula of the form X ~ g where X is a numeric matrix giving the data values and g a factor with two levels giving the corresponding groups.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

...
further arguments to be passed to or from methods.

Details

The classical test for testing the location of a multivariate population or for testing the mean difference for two multivariate populations. When test = "f" the F-distribution is used for the test statistic and it is assumed that the data are normally distributed. If the chisquare approximation is used, the normal assumption can be relaxed to existence of second moments. In the two sample case both populations are assumed to have the same covariance matrix.

The formula interface is only applicable for the 2-sample tests.

Value

A list with class 'htest' containing the following components:

statistic
the value of the T2-statistic. (That is the scaled value of the statistic that has an F distribution or a chisquare distribution depending on the value of test).

parameter
the degrees of freedom for the T2-statistic.

p.value
the p-value for the test.

null.value
the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test.

alternative
a character string with the value 'two.sided'.

method
a character string indicating what type of test was performed.

data.name
a character string giving the name of the data (and grouping vector).

Author(s)

Klaus Nordhausen

References

Anderson, T.W. (2003), An introduction to multivariate analysis, New Jersey: Wiley.

Examples

 
# one sample test:

data(pulmonary)

HotellingsT2(pulmonary) 
HotellingsT2(pulmonary, mu = c(0,0,2), test = "chi")

# two sample test:

set.seed(123456)
X <- rmvnorm(20, c(0, 0, 0, 0), diag(1:4))
Y <- rmvnorm(30, c(0.5, 0.5, 0.5, 0.5), diag(1:4))
Z <- rbind(X, Y)
g <- factor(rep(c(1,2),c(20,30)))

HotellingsT2(X, Y)
HotellingsT2(Z ~ g, mu = rep(-0.5,4))

rm(.Random.seed) 

[Package ICSNP version 1.1-2 Index]