pair.prod {ICSNP}	R Documentation

Pairwise Products

Description

Computes pairwise elementwise products.

Usage

pair.prod(X)

Arguments

X
a numeric matrix.

Details

The function computes all elementwise products of row i and row j with i < j. The function is a wrapper to a C function to do the computation quickly and does no checks concerning the input.

Value

Matrix containing the products.

Author(s)

Klaus Nordhausen

See Also

pair.diff, pair.sum

Examples

X <- matrix(1:10, ncol = 2, byrow = FALSE)
pair.prod(X)

[Package ICSNP version 1.1-2 Index]