pulmonary {ICSNP}	R Documentation

Change in Pulmonary Response after Exposure to Cotton Dust

Description

Changes in pulmonary function of 12 workers after 6 hours of exposure to cotton dust.

Usage

data(pulmonary)

Format

A data frame with 12 observations on the following 3 variables.

FVC

change in FVC (forced vital capacity) after 6 hours.

FEV

change in FEV_3 (forced expiratory volume) after 6 hours.

CC

change in CC (closing capacity) after 6 hours.

Note

There is also a different version of this data set around. In the different version the FVC value of subject 11 is -0.01 instead of -0.10.

Source

Merchant, J. A., Halprin, G. M., Hudson, A. R. Kilburn, K. H., McKenzie, W. N., Hurst, D. J. and Bermazohnm P. (1975), Responses to cotton dust, Archives of Environmental Health, 30, 222–229, Table 5.

Reprinted with permission of the Helen Dwight Reid Educational Foundation. Published by Heldref Publications, 1319 Eighteenth St., NW, Washington, DC 20036-1802.

References

Hettmansperger, T. P. and McKean, J. W. (1998), Robust Nonparametric Statistical Methods, London: Arnold.

Examples

data(pulmonary)
plot(pulmonary)

[Package ICSNP version 1.1-2 Index]