symm.huber {ICSNP}	R Documentation

Symmetrized Huber Scatter Matrix

Description

Iterative algorithm to estimate the symmetrized Huber scatter matrix.

Usage

symm.huber(X, qg = 0.9, init = NULL, eps = 1e-06, maxiter = 100, 
           na.action = na.fail)

Arguments

X
numeric data frame or matrix.

qg
tuning parameter. Should be between 0 and 1. The default is 0.9.

init
an optional matrix giving the starting value for the iteration.

eps
convergence tolerance.

maxiter
maximum number of iterations.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

The symmetrized Huber scatter matrix is the regular Huber scatter matrix for the pairwise differences of the observations taken wrt to the origin.

Note that this function might be memory comsuming and slow for large data sets since the matrix is based on all pairwise difference of the observations.

The function symmhuber in the package SpatialNP offers also a k-step option. The SpatialNP package contains also the function mvhuberM for the regular multivariate Huber location and scatter estimatior.

Value

a matrix.

Author(s)

Klaus Nordhausen and Jari Miettinen

References

Sirkia, S., Taskinen, S. and Oja, H. (2007), Symmetrised M-estimators of scatter. Journal of Multivariate Analysis, 98, 1611–1629.

See Also

symm.huber.wt, symmhuber, mvhuberM

Examples

set.seed(654321)
cov.matrix <- matrix(c(3,2,1,2,4,-0.5,1,-0.5,2), ncol=3)
X <- rmvnorm(100, c(0,0,0), cov.matrix)
symm.huber(X)
rm(.Random.seed)

[Package ICSNP version 1.1-2 Index]