vdw.loc {ICSNP}	R Documentation

Van der Waerden Estimator of Location

Description

Iterative algorithm to compute the location estimator based on van der Waerden scores (sometimes also referred to as normal scores).

Usage

vdw.loc(x, int.diff = 10, maxiter = 1000, na.action = na.fail)

Arguments

x
a numeric vector.

int.diff
number of observations in internal interval when the estimate is searched.

maxiter
maximum number of iterations.

na.action
a function which indicates what should happen when the data contain 'NA's. Default is to fail.

Details

The algorithm searches among the observations and all Walsh averages for the two points nearest around the root of the van der Waerden score criterion. Since the criterion function is monotone first the int.diff of the sorted data points are searched that contain the root. After then determining there the two points of question a linear interpolation is used as an estimate.

Value

the van der Waerden score estimator of location.

Author(s)

Klaus Nordhausen

References

Hettmansperger, T.P. and McKean, J.W. (1998), Robust Nonparametric Statistical Methods, London, Arnold.

Examples

set.seed(1)
x <- rt(100, df = 3)
vdw.loc(x)
rm(.Random.seed)

[Package ICSNP version 1.1-2 Index]